backup/requirements.txt
# Core FastAPI and Web Server
fastapi==0.116.1
uvicorn==0.35.0
python-dotenv==1.1.1
starlette==0.41.2

# HTTP Client
requests==2.32.4
httpx>=0.24.0

# Document Processing Libraries
pypdf==5.8.0
python-docx==1.2.0

# Numerical Computing
numpy==2.2.6

# Vector Database
faiss-cpu==1.11.0.post1

# Embedding Models
sentence-transformers==3.3.1

# LLM Integration (Mistral AI)
mistralai>=1.0.0

# Legacy LLM Integration (Groq) - Keep for fallback
groq>=0.4.1

# Text Processing
langchain-text-splitters==0.3.9

# Additional Utilities
asyncio-throttle>=1.0.0
pydantic>=2.0.0
